+++ libs/components/src/checkbox/checkbox-context.tsx
@@ -3,7 +3,7 @@ import { type Signal, createContextId } from "@builder.io/qwik";
 export const checkboxContextId = createContextId<CheckboxContext>("qds-checkbox-context");
 
 export type CheckboxContext = {
-  isCheckedSig: Signal<boolean | "mixed">;
+  checkedStateSig: Signal<boolean | "mixed">;
   isDisabledSig: Signal<boolean | undefined>;
   isErrorSig: Signal<boolean | undefined>;
   localId: string;
@@ -12,4 +12,8 @@ export type CheckboxContext = {
   required: boolean | undefined;
   value: string | undefined;
   triggerRef: Signal<HTMLButtonElement | undefined>;
+  dataAttributes: Signal<{
+    "data-checked": string | undefined;
+    "data-mixed": string | undefined;
+  }>;
 };