+++ libs/components/src/checkbox/checkbox-hidden-input.tsx
@@ -9,18 +9,18 @@ export const CheckboxHiddenInput = component$(
     // In the case where the native checkbox is checked, but the state is not, we need to update the state
     const handleChange$ = $((e: InputEvent) => {
       const target = e.target as HTMLInputElement;
-      if (target.checked === context.isCheckedSig.value) {
+      if (target.checked === context.checkedStateSig.value) {
         return;
       }
-      context.isCheckedSig.value = target.checked;
+      context.checkedStateSig.value = target.checked;
     });
     return (
       <VisuallyHidden>
         <input
           type="checkbox"
           tabIndex={-1}
-          checked={context.isCheckedSig.value === true}
-          indeterminate={context.isCheckedSig.value === "mixed"}
+          checked={context.checkedStateSig.value === true}
+          indeterminate={context.checkedStateSig.value === "mixed"}
           // Identifier for the hidden native checkbox input element
           data-qds-checkbox-hidden-input
           name={context.name ?? props.name ?? undefined}