+++ libs/components/src/checkbox/checkbox-indicator.tsx
@@ -1,4 +1,11 @@
-import { type PropsOf, Slot, component$, useContext, useStyles$ } from "@builder.io/qwik";
+import {
+  type PropsOf,
+  Slot,
+  component$,
+  useComputed$,
+  useContext,
+  useStyles$
+} from "@builder.io/qwik";
 import { checkboxContextId } from "./checkbox-context";
 import "./checkbox.css";
 import { withAsChild } from "../as-child/as-child";
@@ -9,18 +16,22 @@ export type PublicCheckboxIndicatorProps = PropsOf<"span">;
 export const CheckboxIndicatorBase = component$<PublicCheckboxIndicatorProps>((props) => {
   useStyles$(styles);
   const context = useContext(checkboxContextId);
+
+  const isHiddenSig = useComputed$(() => context.checkedStateSig.value === false);
+  const isCheckedSig = useComputed$(() => context.checkedStateSig.value === true);
+
   return (
     <Render
       {...props}
       fallback="span"
       // Indicates whether the indicator should be hidden based on checkbox state
-      data-hidden={!context.isCheckedSig.value}
+      data-hidden={isHiddenSig.value}
       // Indicates whether the checkbox is in a checked state
-      data-checked={context.isCheckedSig.value ? "" : undefined}
+      data-checked={isCheckedSig.value}
       // Indicates whether the checkbox is in an indeterminate state
-      data-mixed={context.isCheckedSig.value === "mixed" ? "" : undefined}
+      data-mixed={context.checkedStateSig.value === "mixed" ? "" : undefined}
       // Identifier for the checkbox indicator element
-      data-qds-indicator
+      data-qds-checkbox-indicator
       aria-hidden="true"
     >
       <Slot />