+++ libs/components/src/checklist/checklist-item-indicator.tsx
@@ -1,12 +1,12 @@
 import { type PropsOf, Slot, component$ } from "@builder.io/qwik";
-import { Checkbox } from "..";
+import { CheckboxIndicatorBase } from "../checkbox/checkbox-indicator";
 
 export const ChecklistItemIndicator = component$(
-  (props: PropsOf<typeof Checkbox.Indicator>) => {
+  (props: PropsOf<typeof CheckboxIndicatorBase>) => {
     return (
-      <Checkbox.Indicator {...props}>
+      <CheckboxIndicatorBase {...props}>
         <Slot />
-      </Checkbox.Indicator>
+      </CheckboxIndicatorBase>
     );
   }
 );