+++ libs/components/src/radio-group/radio-group-trigger.tsx
@@ -12,6 +12,7 @@ import { getNextIndex } from "../../utils/indexer";
 import { withAsChild } from "../as-child/as-child";
 import { Render } from "../render/render";
 import { radioGroupContextId } from "./radio-group-context";
+import { radioGroupItemContextId } from "./radio-group-item";
 
 type PublicTriggerProps = PropsOf<"button"> & {
   _index?: number;
@@ -19,9 +20,10 @@ type PublicTriggerProps = PropsOf<"button"> & {
 
 export const RadioGroupTriggerBase = component$((props: PublicTriggerProps) => {
   const context = useContext(radioGroupContextId);
+  const itemContext = useContext(radioGroupItemContextId);
   const triggerRef = useSignal<HTMLElement>();
   const { _index, ...restProps } = props;
-  const value = context.itemValue;
+  const value = itemContext.itemValue;
 
   useTask$(function getIndexOrder() {
     if (_index === undefined) {
@@ -34,18 +36,17 @@ export const RadioGroupTriggerBase = component$((props: PublicTriggerProps) => {
     };
   });
 
-  const isSelectedSig = useComputed$(() => context.selectedValueSig.value === value);
   const isDisabledSig = useComputed$(() => context.isDisabledSig.value || props.disabled);
 
   const tabIndexSig = useComputed$(() => {
-    const isSelected = isSelectedSig.value;
+    const isSelected = itemContext.isSelectedSig.value;
     const isFirstItem = _index === 0;
     const noSelection = !context.selectedValueSig.value;
 
     return isSelected || (noSelection && isFirstItem) ? 0 : -1;
   });
 
-  const handleClick$ = $(() => {
+  const handleSelection$ = $(() => {
     if (isDisabledSig.value) return;
     context.selectedValueSig.value = value;
   });
@@ -54,7 +55,7 @@ export const RadioGroupTriggerBase = component$((props: PublicTriggerProps) => {
     if (isDisabledSig.value) return;
 
     if (event.key === " " || event.key === "Enter") {
-      await handleClick$();
+      await handleSelection$();
     }
   });
 
@@ -65,12 +66,12 @@ export const RadioGroupTriggerBase = component$((props: PublicTriggerProps) => {
       ref={triggerRef}
       type="button"
       role="radio"
-      aria-checked={isSelectedSig.value}
-      data-checked={isSelectedSig.value}
+      aria-checked={itemContext.isSelectedSig.value}
+      data-checked={itemContext.isSelectedSig.value}
       data-qds-radio-group-trigger
       data-disabled={isDisabledSig.value || undefined}
       value={value}
-      onClick$={[handleClick$, props.onClick$]}
+      onClick$={[handleSelection$, props.onClick$]}
       onKeyDown$={[handleKeyDown$, props.onKeyDown$]}
       disabled={isDisabledSig.value}
       tabIndex={tabIndexSig.value}